# swagger/momentsz4ever-api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Momentz-4-ever Services",
    "description": "Momentz-4-ever API specs",
    "termsOfService": ""
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
       "get": {
          "tags": ["Users"],
          "description": "Gets all the Users of the application",
          "operationId": "get",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "default":{
              "description":"Bad Request"
            },
            "200": {
              "description": "returns an array of users",
              "schema": {
                "type" : "array",
               "items": {
                 "$ref": "#/definitions/Users"
              } 
            }
          }
        }
       },
         "post":{
          "tags":["Users"],
          "description":"Adds an User",
          "operationId":"post",
          "produces": [
            "application/json"
          ],
          "parameters":[
              {
                "name": "body",
                "in": "body",
                "description": "User object to be updated",
                "required": true,
                "schema": {
                     "$ref": "#/definitions/Users"
                }
              }
          ],
          "responses": {
             "default":{
              "description":"Bad Request"
            },
            "201": {
              "description": "Successfully created",
              "schema":{
                "$ref":"#/definitions/Users"
              }
            }
          }
    }
    },
    "/albums": {
       "get": {
          "tags": ["Albums"],
          "description": "Gets all the Albums of the application",
          "operationId": "get",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
             "default":{
              "description":"Bad Request"
            },
            "200": {
              "description": "returns an array of albums",
              "schema": {
                 "type" : "array",
               "items": {
                 "$ref": "#/definitions/Albums"
              } 
              }
            }
          }
        },
      "post":{
          "tags":["Albums"],
          "description":"Adds an album",
          "operationId":"post",
          "produces": [
            "application/json"
          ],
          "parameters":[
              {
                "name": "body",
                "in": "body",
                "description": "Albums object that needs to be updated",
                "required": true,
                "schema": {
                     "$ref": "#/definitions/Albums"
                }
              }       
          ],
          "responses": {
             "default":{
              "description":"Bad Request"
            },
            "201": {
              "description": "Successfully created",
              "schema":{
                "$ref":"#/definitions/Albums"
              }
            }
          }
        }
  },
    "/albums/{id}": {
          "put":{
          "tags":["Albums by ID"],
          "description":"Updates an album",
          "operationId":"put",
          "produces": [
            "application/json"
          ],
          "parameters":[
                {
                    "name": "albumId",
                    "in": "path",
                    "description": "Album id that needs to be updated",
                    "required": true,
                    "type": "string"
                },
                {
                  "name": "body",
                  "in": "body",
                  "description": "Album object that needs to be updated",
                  "required": true,
                  "schema": {
                     "$ref": "#/definitions/Albums"
                }
              }   
            ],
          "responses": {
             "201": {
              "description": "Successfully created",
              "schema":{
                "$ref":"#/definitions/Albums"
              }
            }
          }
        },
        "delete":{
          "tags":["Albums by ID"],
          "description":"Deletes an album",
          "operationId":"delete",
          "produces": [
            "application/json"
          ],
          "parameters":[
              {
                "name": "albumId",
                "in": "path",
                "description": "Album-Id to delete",
                "required": true,
                "type": "string"
              }
            ],
          "responses": {
            "200": {
              "description": "Successful Operation",
              "schema":{
                "$ref":"#/definitions/Albums"
              }
            }
          }
        },
      "get":{
          "tags": [ "Albums by ID"],
          "description": "Returns a single album",
          "operationId": "findById",
          "produces": [
            "application/json"
          ],
          "parameters": [
              {
                "name": "albumId",
                "in": "path",
                "description": "Id of Album to return",
                "required": true,
                "type": "string"
              }
            ],
        "responses": {
           "200": {
              "description": "Successful Operation",
              "schema": {
                "$ref": "#/definitions/Albums"
              }
            }
          }
        }
      },
  "/users/{id}": {
        "put":{
          "tags":["Users by ID"],
          "description":"Updates an User",
          "operationId":"put",
          "produces": [
            "application/json"
          ],
          "parameters":[
                {
                    "name": "id",
                    "in": "path",
                    "description": "USer id that needs to be updated",
                    "required": true,
                    "type": "string"
                },
                {
                  "name": "body",
                  "in": "body",
                  "description": "User object that needs to be updated",
                  "required": true,
                  "schema": {
                     "$ref": "#/definitions/Users"
                }
              }   
            ],
          "responses": {
             "201": {
              "description": "Successfully created",
              "schema":{
                "$ref":"#/definitions/Users"
              }
            }
          }
        },
        "delete":{
          "tags":["Users by ID"],
          "description":"Deletes a User",
          "operationId":"delete",
          "produces": [
            "application/json"
          ],
          "parameters":[
              {
                "name": "id",
                "in": "path",
                "description": "User-Id to delete",
                "required": true,
                "type": "string"
              }
            ],
          "responses": {
             "200": {
              "description": "Successful Operation",
              "schema":{
                "$ref":"#/definitions/Users"
              }
            }
          }
        },
      "get":{
          "tags": [ "Users by ID"],
          "description": "Returns a User",
          "operationId": "findById",
          "produces": [
            "application/json"
          ],
          "parameters": [
              {
                "name": "id",
                "in": "path",
                "description": "Id of User to return",
                "required": true,
                "type": "string"
              }
            ],
        "responses": {
           "200": {
              "description": "Successful Operation",
              "schema": {
                "$ref": "#/definitions/Users"
              }
            }
          }
        }
      }
  },
  "definitions": {
    "Users": {
      "properties": {
         "id": {
          "type": "integer",
          "example": "0"
        },
         "firstName":{
           "type": "string",
           "example": "James"
        },
         "lastName":{
           "type": "string",
           "example": "Turner"
        },
         "email":{
           "type": "string",
           "example": "Turner@gmail.com"
        }
      }
    },

    "Albums": {
        "properties": {
        "title":{
          "type": "string",
          "example": "Album 1"
        },
        "albumId": {
          "type": "integer",
          "example": "0"
        },
        "description":{
          "type": "string",
          "example": "This is the first Album"
        },
        "imageUrl":{
          "type": "string",
          "example": "https://nodejs.org/static/images/logos/nodejs-new-pantone-black.png"
        },
        "createdDate":{
          "type": "string",
          "example": "05/07/2017"
        },
        "lastViewedOn":{
          "type": "string",
          "example": "05/08/2017"
        }
        }
    }
   }
}